+++ doc/ref/RefDocExamples/RefDocExamples/main.m
@@ -8,7 +8,7 @@ int main(int argc, char * argv[])
     @autoreleasepool {
         ex_objc_query_dynamic_intro();
         ex_objc_query_typed_intro();
-        ex_objc_context_intro();
+        ex_objc_transaction_manager_intro();
         ex_objc_realm_intro();
         ex_objc_table_dynamic_intro();
         ex_objc_table_typed_intro();