+++ src/realm/objc/RLMConstants.h
@@ -25,4 +25,4 @@
  * Posted by RLMRealm when it changes, that is when a table is
  * added, removed, or changed in any way.
  */
-extern NSString *const RLMContextDidChangeNotification;
+extern NSString *const RLMRealmDidChangeNotification;